+++ src/main/java/com/alterra/capstone/repository/InstructorRepository.java
@@ -4,6 +4,6 @@ import com.alterra.capstone.entity.Instructor;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface InstructorRepository extends JpaRepository<Instructor, Integer> {
+public interface InstructorRepository extends JpaRepository<Instructor, Long> {
 
 }
