+++ src/main/java/com/alterra/capstone/service/RoleService.java
@@ -5,9 +5,9 @@ import com.alterra.capstone.payload.RolePayload;
 
 public interface RoleService {
     Iterable<Role> getAll();
-    Role getById (Integer id);
+    Role getById (Long id);
     Role create (RolePayload rolePayload);
-    void delete(Integer id);
+    void delete(Long id);
 
 
 }
